+++ RxBlocking/BlockingObservable+Operators.swift
@@ -10,6 +10,19 @@
     import RxSwift
 #endif
 
+/// The `MaterializedSequenceResult` enum represents the materialized
+/// output of a BlockingObservable.
+///
+/// If the sequence terminates successfully, the result is represented
+/// by `.completed` with the array of elements.
+///
+/// If the sequence terminates with error, the result is represented
+/// by `.failed` with both the array of elements and the terminating error.
+public enum MaterializedSequenceResult<T> {
+    case completed(elements: [T])
+    case failed(elements: [T], error: Error)
+}
+
 extension BlockingObservable {
     /// Blocks current thread until sequence terminates.
     ///
@@ -17,7 +30,8 @@ extension BlockingObservable {
     ///
     /// - returns: All elements of sequence.
     public func toArray() throws -> [E] {
-        return try convertToArray()
+        let results = materializeResult()
+        return try elementsOrThrow(results)
     }
 }
 
@@ -28,7 +42,8 @@ extension BlockingObservable {
     ///
     /// - returns: First element of sequence. If sequence is empty `nil` is returned.
     public func first() throws -> E? {
-        return try convertToArray(max: 1).first
+        let results = materializeResult(max: 1)
+        return try elementsOrThrow(results).first
     }
 }
 
@@ -39,7 +54,8 @@ extension BlockingObservable {
     ///
     /// - returns: Last element in the sequence. If sequence is empty `nil` is returned.
     public func last() throws -> E? {
-        return try convertToArray().last
+        let results = materializeResult()
+        return try elementsOrThrow(results).last
     }
 }
 
@@ -60,7 +76,8 @@ extension BlockingObservable {
     /// - parameter predicate: A function to test each source element for a condition.
     /// - returns: Returns the only element of an sequence that satisfies the condition in the predicate, and reports an error if there is not exactly one element in the sequence.
     public func single(_ predicate: @escaping (E) throws -> Bool) throws -> E? {
-        let elements = try convertToArray(max: 2, predicate: predicate)
+        let results = materializeResult(max: 2, predicate: predicate)
+        let elements = try elementsOrThrow(results)
         
         switch elements.count {
         case 0:
@@ -74,9 +91,19 @@ extension BlockingObservable {
 }
 
 extension BlockingObservable {
-    fileprivate func convertToArray(max: Int? = nil, predicate: @escaping (E) throws -> Bool = { _ in true }) throws -> [E] {
+    /// Blocks current thread until sequence terminates.
+    ///
+    /// The sequence is materialized as a result type capturing how the sequence terminated (completed or error), along with any elements up to that point.
+    ///
+    /// - returns: On completion, returns the list of elements in the sequence. On error, returns the list of elements up to that point, along with the error itself.
+    public func materialize() -> MaterializedSequenceResult<E> {
+        return materializeResult()
+    }
+}
+
+extension BlockingObservable {
+    fileprivate func materializeResult(max: Int? = nil, predicate: @escaping (E) throws -> Bool = { _ in true }) -> MaterializedSequenceResult<E> {
         var elements: [E] = Array<E>()
-        
         var error: Swift.Error?
         
         let lock = RunLoopLock(timeout: timeout)
@@ -127,9 +154,18 @@ extension BlockingObservable {
         }
         
         if let error = error {
-            throw error
+            return MaterializedSequenceResult.failed(elements: elements, error: error)
         }
         
-        return elements
+        return MaterializedSequenceResult.completed(elements: elements)
+    }
+    
+    fileprivate func elementsOrThrow(_ results: MaterializedSequenceResult<E>) throws -> [E] {
+        switch results {
+        case .failed(_, let error):
+            throw error
+        case .completed(let elements):
+            return elements
+        }
     }
 }